chore/style: apply Clippy and Rustfmt
