+++ src/mock_crust_detail/test_node.rs
@@ -15,21 +15,21 @@
 // Please review the Licences for the specific language governing permissions and limitations
 // relating to use of the SAFE Network Software.
 
-use std::fs;
-use std::path::PathBuf;
 
 use config_handler::Config;
 use kademlia_routing_table::RoutingTable;
-use rand::{self, Rng};
-use routing::mock_crust::{self, Endpoint, Network, ServiceHandle};
-use routing::XorName;
-use rustc_serialize::hex::ToHex;
-use std::env;
-use vault::Vault;
 
 use personas::data_manager::IdAndVersion;
+use rand::{self, Rng};
+use routing::XorName;
+use routing::mock_crust::{self, Endpoint, Network, ServiceHandle};
+use rustc_serialize::hex::ToHex;
+use std::env;
+use std::fs;
+use std::path::PathBuf;
 
 use super::poll;
+use vault::Vault;
 
 /// Test node for mock network
 pub struct TestNode {
