+++ tests/maid_manager.rs
@@ -22,8 +22,8 @@ use itertools::Itertools;
 use rand::Rng;
 use rand::distributions::{IndependentSample, Range};
 use routing::{Data, GROUP_SIZE, ImmutableData, XorName};
-use routing::mock_crust::{self, Network};
 use routing::client_errors::{GetError, MutationError};
+use routing::mock_crust::{self, Network};
 use safe_vault::Config;
 use safe_vault::mock_crust_detail::{self, poll, test_node};
 use safe_vault::mock_crust_detail::test_client::TestClient;
